docs/sample-apps/jfk-files/requirements.txt
beautifulsoup4
requests
mistralai
pixeltable
sentence-transformers
tiktoken
spacy
mcp
httpx
uvicorn
starlette
python-dotenv